user-service/api/requirements.txt
fastapi
uvicorn[stardard]
sqlalchemy
pymysql
python-multipart
passlib[bcrypt]
python-jose[cryptography]
bcrypt
PyJWT